New feature: synergy (#166)
* update creature table: add race and class

* add synergy logic

* implement Orc synergy: add max health

* refactor simulator and make it a system

* more synergy but without testing

* test and fix synergies

* add race synergy of troll

* chore

* update human synergy

* update hero refreshing

* add sub-systems back into config

+++ packages/contracts/src/systems/MatchingSystem.sol
@@ -160,7 +160,7 @@ contract MatchingSystem is System {
         // }
 
         /// @dev initalize inventory
-        uint16[] memory inventory = new uint16[](GameConfig.getInventorySlotNum(0));
+        uint24[] memory inventory = new uint24[](GameConfig.getInventorySlotNum(0));
 
         uint256 num = _players.length;
         for (uint256 i; i < num; ++i) {
